\name{internal functions}

\alias{bg.correct.shift}
\alias{bg.correct.stretch}
\alias{get.background.code}
\alias{get.normalization.code}
\alias{get.psi.code}
\alias{get.summary.code}
\alias{get.default.psi.k}
\alias{convert.LESN.param}
\alias{PLM.designmatrix}
\alias{PLM.designmatrix2}

\title{Internal AffyExtensions functions}

\description{Internal AffyExtensions functions}

\details{These are not to be called by the user and/or are undergoing testing}

\keyword{internal}
