\name{internal functions}

\alias{bg.correct.shift}
\alias{bg.correct.stretch}
\alias{bg.correct.gcrma}
\alias{get.background.code}
\alias{get.normalization.code}
\alias{get.psi.code}
\alias{get.summary.code}
\alias{get.default.psi.k}
\alias{convert.LESN.param}
\alias{PLM.designmatrix}
\alias{PLM.designmatrix2}
\alias{fitPLM.old}
\alias{PLM.designmatrix3}
\alias{verify.bg.param}
\alias{verify.constraint.types}
\alias{verify.model.param}
\alias{verify.norm.param}
\alias{verify.output.param}
\alias{verify.variable.types}

\alias{normalize.AffyBatch.quantiles.chromosome}


\title{Internal affyPLM functions}

\description{Internal affyPLM functions}

\details{These are not to be called by the user and/or are undergoing testing}

\keyword{internal}
